requirements.txt
flask
flask_socketio
zenora
python-dotenv
werkzeug
mysql-connector-python
